+++ src/components/App/App.tsx
@@ -2,9 +2,9 @@ import './App.css';
 import React, { FC } from 'react';
 import { Switch, Route, Link } from 'react-router-dom';
 import { useSelector } from 'react-redux';
-import { getTest } from '../../selectors';
-import { appActions } from '../../reducer';
-import { ROUTES } from '../../constants';
+import { getTest } from '../../store/selectors';
+import { appActions } from '../../store/reducer';
+import { ROUTES } from '../../common/constants';
 import { SignIn } from '../../features/signin/SignIn';
 import { Forum } from '../../features/forum/Forum';
 import { SignUp } from '../../features/signup/SignUp';
