+++ src/features/account/selectors.ts
@@ -1,4 +1,4 @@
-import { RootState } from '../../store';
+import { RootState } from '../../store/store';
 
 const getAccountState = (state: RootState) => state.account;
